refactor: rename and structure components files
