+++ src/app/rest-client/components/response-section.tsx
